backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pymongo==4.6.0
requests==2.31.0
openai==1.3.0
anthropic==0.7.0
uuid==1.30
websockets==12.0
aiofiles==23.2.1